+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Upgrade quran search engine package to the athar version
+
 ## [5.0.0-athar] - 18-03-2026
 
 ## Added